only process Sylius Entities in metadata subscribers
