+++ src/Sylius/Bundle/ResourceBundle/Resources/config/services.xml
@@ -33,8 +33,8 @@
 
         <parameter key="sylius.event_subscriber.resource_delete.class">Sylius\Bundle\ResourceBundle\EventListener\ResourceDeleteSubscriber</parameter>
 
-        <parameter key="sylius.event_subscriber.load_orm_metadata.class">Sylius\Bundle\ResourceBundle\EventListener\LoadORMMetadataSubscriber</parameter>
-        <parameter key="sylius.event_subscriber.load_odm_metadata.class">Sylius\Bundle\ResourceBundle\EventListener\LoadODMMetadataSubscriber</parameter>
+        <parameter key="sylius.event_subscriber.orm_mapped_super_class.class">Sylius\Bundle\ResourceBundle\EventListener\ORMMappedSuperClassSubscriber</parameter>
+        <parameter key="sylius.event_subscriber.odm_mapped_super_class.class">Sylius\Bundle\ResourceBundle\EventListener\ODMMappedSuperClassSubscriber</parameter>
 
         <parameter key="sylius.validator.enabled.class">Sylius\Bundle\ResourceBundle\Validator\EnabledValidator</parameter>
         <parameter key="sylius.validator.disabled.class">Sylius\Bundle\ResourceBundle\Validator\DisabledValidator</parameter>
@@ -65,11 +65,11 @@
             <tag name="kernel.event_subscriber" event="kernel.exception" />
         </service>
 
-        <service id="sylius.event_subscriber.load_orm_metadata" class="%sylius.event_subscriber.load_orm_metadata.class%">
+        <service id="sylius.event_subscriber.orm_mapped_super_class" class="%sylius.event_subscriber.orm_mapped_super_class.class%">
             <argument type="service" id="sylius.resource_registry" />
             <tag name="doctrine.event_subscriber" priority="8192" />
         </service>
-        <service id="sylius.event_subscriber.load_odm_metadata" class="%sylius.event_subscriber.load_odm_metadata.class%">
+        <service id="sylius.event_subscriber.odm_mapped_super_class" class="%sylius.event_subscriber.odm_mapped_super_class.class%">
             <argument type="service" id="sylius.resource_registry" />
             <tag name="doctrine_mongodb.odm.event_subscriber" priority="8192" />
         </service>
